+++ CHANGELOG.md
@@ -116,6 +116,7 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
 * Roofline support for Ubuntu 20.04 and SLES below 15.6
 * Usage of rocm-smi
 * Remove support for MI50/MI60 in accordance with the documentation
+* Hardware IP block based filtering has been removed in favor of analysis report block based filtering
 
 ## ROCm Compute Profiler 3.1.0 for ROCm 6.4.0